+++ cypress/integration/2.10-fullRefund.spec.js
@@ -1,4 +1,8 @@
-import { requestRefund, singleProduct } from '../support/outputvalidation'
+import {
+  requestRefund,
+  singleProduct,
+  externalSeller,
+} from '../support/outputvalidation'
 import { loginViaCookies } from '../support/common/support.js'
 import { refund } from '../support/common/refund_apis.js'
 import { getRefundPayload } from '../support/refund_payload.js'
@@ -9,7 +13,7 @@ describe('Testing Cybersource transaction API for full refund', () => {
   // Load test setup
   loginViaCookies()
 
-  const { prefix } = singleProduct
+  const { prefix, totalAmount } = singleProduct
   const { transactionIdEnv, paymentTransactionIdEnv } = getTestVariables(prefix)
 
   it('Verify whether we have an order to request for full refund', () => {
@@ -23,16 +27,19 @@ describe('Testing Cybersource transaction API for full refund', () => {
   // Request full refund for the ordered product added in 2.1-singleProduct.spec.js
   refund(
     {
-      total: requestRefund.getFullRefundTotal, // Amount
+      total: totalAmount, // Amount
       title: 'full', // Refund Type for test case title
       env: requestRefund.fullRefundEnv, // variable name where we stored the orderid in node environment
+      externalSeller,
     },
-    getRefundPayload
+    getRefundPayload,
+    { sendInvoice: true, startHandling: true }
   )
 
   // verify cybersource transaction
   verifyRefundTid({
     prefix: 'fullRefund',
     paymentTransactionIdEnv,
+    payerAuth: true,
   })
 })