+++ cypress/integration/2.9-settlements.spec.js
@@ -1,5 +1,4 @@
 import {
-  verifyPaymentSettled,
   paymentTestCases,
   APITestCases,
   verifyCyberSourceAPI,
@@ -21,7 +20,7 @@ describe('Verify settlements for ordered products', () => {
 
   // SingleProduct API testcase
   APITestCases(
-    { prefix: singleProduct.prefix, approved: true },
+    { prefix: singleProduct.prefix, approved: true, referenceSuffix: true },
     {
       ...getTestVariables(singleProduct.prefix),
       orderIdEnv: requestRefund.fullRefundEnv,
@@ -39,7 +38,7 @@ describe('Verify settlements for ordered products', () => {
 
   // discountProduct
   APITestCases(
-    { prefix: discountProduct.prefix, approved: true },
+    { prefix: discountProduct.prefix, approved: true, referenceSuffix: true },
     {
       ...getTestVariables(discountProduct.prefix),
       orderIdEnv: discountProduct.env,
@@ -65,6 +64,7 @@ describe('Verify settlements for ordered products', () => {
     transactionIdEnv,
     paymentTransactionIdEnv,
     approved: true,
+    orderIdEnv: externalSeller.directSaleEnv,
   })
 
   paymentTestCases(
@@ -82,5 +82,4 @@ describe('Verify settlements for ordered products', () => {
       orderIdEnv: promotionProduct.env,
     }
   )
-  verifyPaymentSettled(promotionProduct.prefix, promotionProduct.env)
 })